# performance: replace a critical usage of Array.concat with for loop/push