--- a/lib/model.js
+++ b/lib/model.js
@@ -1578,7 +1578,10 @@ class Model {
             // Force array so we can concat no matter if it's 1:1 or :M
             if (!Array.isArray(associations)) associations = [associations];
 
-            return memo.concat(associations);
+            for (let i = 0, len = associations.length; i !== len; ++i) {
+              memo.push(associations[i]);
+            }
+            return memo;
           }, []),
           _.assign(
             {},