@sushantdhiman 2017-02-03
@natesilva Could you please compare results using https://github.com/sushantdhiman/sequelize-benchmark , It will help us decide if there is any significant performance improvement and how much.