@natesilva 2017-02-13
I haven’t been able to run the benchmark. With MySQL it reports a `Got a packet bigger than 'max_allowed_packet' bytes`; with PostgreSQL it never seems to finish. (This isn’t a problem in my PR, it does this even against the Sequelize master branch.) So my next step is to debug the benchmark.